Split cell tests
